code refactoring part 1
tasks decomposition
